pyproject.toml
[tool.poetry]
name = "bitcoinprice"
version = "0.1.0"
description = ""
authors = ["Rafael Alvarenga <devrafaelalvarenga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
